chore: export from root package for non-subpath builds
