+++ browser/chrome-extension/tsconfig.json
@@ -11,16 +11,6 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "baseUrl": "./",
-    "paths": {
-      "@mdip/cipher/web": ["../../packages/cipher/src/cipher-web.ts"],
-      "@mdip/cipher/types": ["../../packages/cipher/dist/types/types.d.ts"],
-      "@mdip/common/errors": ["../../packages/common/src/errors.ts"],
-      "@mdip/gatekeeper/types": ["../../packages/gatekeeper/src/types.ts"],
-      "@mdip/gatekeeper/client": ["../../packages/gatekeeper/src/gatekeeper-client.ts"],
-      "@mdip/keymaster": ["../../packages/keymaster/src/keymaster.ts"],
-      "@mdip/keymaster/types": ["../../packages/keymaster/src/types.ts"],
-      "@mdip/keymaster/wallet/*": ["../../packages/keymaster/src/db/*"]
-    }
+    "baseUrl": "./"
   }
 }
