+++ demo/commonjs-demo/sample.env
@@ -1,0 +1,2 @@
+#GATEKEEPER_SERVICE_URL="http://localhost:4224"
+#KEYMASTER_SERVICE_URL="http://localhost:4226"
