+++ jest.config.js
@@ -16,7 +16,6 @@ const config = {
         '^@mdip/cipher/types': '<rootDir>/packages/cipher/src/types.ts',
         '^@mdip/common/errors$': '<rootDir>/packages/common/src/errors.ts',
         '^@mdip/common/utils$': '<rootDir>/packages/common/src/utils.ts',
-        '^@mdip/cipher$': '<rootDir>/packages/cipher/src/cipher.ts',
         '^@mdip/gatekeeper$': '<rootDir>/packages/gatekeeper/src/gatekeeper.ts',
         '^@mdip/gatekeeper/types$': '<rootDir>/packages/gatekeeper/src/types.ts',
         '^@mdip/gatekeeper/client$': '<rootDir>/packages/gatekeeper/src/gatekeeper-client.ts',
