+++ packages/gatekeeper/package.json
@@ -3,17 +3,27 @@
   "version": "1.1.1",
   "description": "MDIP Gatekeeper",
   "type": "module",
-  "module": "./dist/esm/gatekeeper.js",
-  "main": "./dist/cjs/gatekeeper.cjs",
-  "types": "./dist/types/gatekeeper.d.ts",
+  "module": "./dist/esm/node.js",
+  "main": "./dist/cjs/node.cjs",
+  "types": "./dist/types/node.d.ts",
   "files": [
     "dist"
   ],
   "exports": {
     ".": {
-      "import": "./dist/esm/gatekeeper.js",
-      "require": "./dist/cjs/gatekeeper.cjs",
-      "types": "./dist/types/gatekeeper.d.ts"
+      "browser": {
+        "import": "./dist/esm/index.js",
+        "types": "./dist/types/index.d.ts"
+      },
+      "node": {
+        "import": "./dist/esm/node.js",
+        "require": "./dist/cjs/node.cjs",
+        "types": "./dist/types/node.d.ts"
+      },
+      "default": {
+        "import": "./dist/esm/node.js",
+        "types": "./dist/types/node.d.ts"
+      }
     },
     "./client": {
       "import": "./dist/esm/gatekeeper-client.js",
