+++ packages/gatekeeper/src/gatekeeper-client.ts
@@ -2,13 +2,14 @@ import axios, { AxiosError } from 'axios';
 import {
     BlockId,
     BlockInfo,
+    GatekeeperClientOptions,
     GatekeeperInterface,
     GatekeeperEvent,
+    GetStatusResult,
     Operation,
     MdipDocument,
     ResolveDIDOptions,
     GetDIDOptions,
-    CheckDIDsResult,
     ImportBatchResult,
     ProcessEventsResult,
     VerifyDbResult,
@@ -21,22 +22,6 @@ function throwError(error: AxiosError | any): never {
         throw error.response.data;
     }
     throw error.message;
-}
-
-export interface GatekeeperClientOptions {
-    url?: string;
-    console?: typeof console;
-    waitUntilReady?: boolean;
-    intervalSeconds?: number;
-    chatty?: boolean;
-    becomeChattyAfter?: number;
-    maxRetries?: number;
-}
-
-export interface GetStatusResult {
-    uptimeSeconds: number;
-    dids: CheckDIDsResult;
-    memoryUsage: NodeJS.MemoryUsage;
 }
 
 export default class GatekeeperClient implements GatekeeperInterface {
