+++ packages/gatekeeper/src/index.ts
@@ -1,0 +1,3 @@
+export { default as GatekeeperClient } from './gatekeeper-client.js';
+export { default as DbJsonMemory } from './db/json-memory.js';
+export * from './types.js';
