+++ packages/ipfs/package.json
@@ -3,10 +3,28 @@
   "version": "1.1.1",
   "description": "MDIP IPFS lib",
   "type": "module",
+  "module": "./dist/esm/node.js",
+  "main": "./dist/cjs/node.cjs",
+  "types": "./dist/types/node.d.ts",
   "files": [
     "dist"
   ],
   "exports": {
+    ".": {
+      "browser": {
+        "import": "./dist/esm/index.js",
+        "types": "./dist/types/index.d.ts"
+      },
+      "node": {
+        "import": "./dist/esm/node.js",
+        "require": "./dist/cjs/node.cjs",
+        "types": "./dist/types/node.d.ts"
+      },
+      "default": {
+        "import": "./dist/esm/node.js",
+        "types": "./dist/types/node.d.ts"
+      }
+    },
     "./helia": {
       "import": "./dist/esm/helia-client.js",
       "require": "./dist/cjs/helia-client.cjs",
