+++ packages/ipfs/src/index.ts
@@ -1,0 +1,3 @@
+export { default as KuboClient } from './kubo-client.js';
+export * from './utils.js';
+export * from './types.js';
