+++ packages/ipfs/src/node.ts
@@ -1,0 +1,2 @@
+export * from './index.js';
+export { default as HeliaClient } from './helia-client.js';
