+++ packages/keymaster/src/index.ts
@@ -1,0 +1,9 @@
+export { default } from './keymaster.js';
+export { default as KeymasterClient } from './keymaster-client.js';
+export { default as WalletJsonMemory } from './db/json-memory.js';
+export { default as WalletWeb } from './db/web.js';
+export { default as WalletWebEncrypted } from './db/web-enc.js';
+export { default as WalletCache } from './db/cache.js';
+export { default as WalletChrome } from './db/chrome.js';
+export * from './db/typeGuards.js';
+export * from './types.js';
