+++ packages/keymaster/src/node.ts
@@ -1,0 +1,7 @@
+export * from './index.js';
+export { default } from './index.js';
+export { default as WalletJson } from './db/json.js';
+export { default as WalletEncrypted } from './db/json-enc.js';
+export { default as WalletRedis } from './db/redis.js';
+export { default as WalletMongo } from './db/mongo.js';
+export { default as WalletSQLite } from './db/sqlite.js';
